requirements.txt
alethic-ism-core==1.0.24
alethic-ism-db==1.0.26
pyyaml
nats-py
python-dotenv
pydantic
psycopg2
openai
tiktoken
mako